# Vinayak/gemm kernel