Changes:

1) Added support for scaling 8-bit dtypes. This is the typical use case in workloads
2) Cleaned up the script and removed references / comments from 03 tutorial
3) Other code cleanup
4) Added new autotune config (provided by @zhanglx13).

Performance comparison (speedup is after / before)

16-bit

| M    | N    | K     | Before.Fp16 | Before.Bf16 | After.Fp16 | After.Bf16 | Speedup fp16 | Speedup bf16  |
|------|------|-------|-------------|-------------|-------------|-------------|--------------|---------------|
| 1024 | 1024 | 1024  | 51          | 51          | 47          | 47          | 0.92         | 0.93          |
| 2048 | 2048 | 2048  | 232         | 237         | 237         | 240         | 1.02         | 1.01          |
| 3072 | 3072 | 3072  | 359         | 382         | 341         | 346         | 0.95         | 0.91          |
| 4096 | 4096 | 4096  | 356         | 376         | 496         | 512         | 1.39         | 1.36          |
| 5120 | 5120 | 5120  | 353         | 365         | 395         | 412         | 1.12         | 1.13          |
| 6144 | 6144 | 6144  | 351         | 359         | 472         | 505         | 1.34         | 1.41          |
| 7168 | 7168 | 7168  | 344         | 358         | 504         | 493         | 1.47         | 1.38          |
| 8192 | 8192 | 8192  | 362         | 390         | 468         | 498         | 1.29         | 1.28          |
| 4864 | 4096 | 8192  | 381         | 401         | 553         | 579         | 1.45         | 1.44          |
| 9728 | 8192 | 65536 | 247         | 263         | 539         | 518         | 2.18         | 1.97          |
| 4864 | 8192 | 4160  |             |             | 555         | 587         | NA           | NA            |


8-bit

| M    | N    | K     | Before.Int8 | Before.Fp8E4 | Before.Fp8E5 | After.Int8 | After.Fp8E4 | After.Fp8E5 | Speedup int8 | Speedup fp8 | Speedup bf8  |
|------|------|-------|-------------|--------------|--------------|-------------|--------------|--------------|--------------|-------------|--------------|
| 1024 | 1024 | 1024  | 51          | 50           | 49           | 47          | 45           | 45           | 0.92         | 0.90        | 0.91         |
| 2048 | 2048 | 2048  | 338         | 293          | 288          | 322         | 325          | 329          | 0.95         | 1.11        | 1.14         |
| 3072 | 3072 | 3072  | 594         | 577          | 555          | 562         | 562          | 566          | 0.95         | 0.97        | 1.02         |
| 4096 | 4096 | 4096  | 921         | 916          | 803          | 813         | 818          | 830          | 0.88         | 0.89        | 1.03         |
| 5120 | 5120 | 5120  | 737         | 709          | 663          | 717         | 711          | 726          | 0.97         | 1.00        | 1.09         |
| 6144 | 6144 | 6144  | 1059        | 980          | 911          | 909         | 908          | 919          | 0.86         | 0.93        | 1.01         |
| 7168 | 7168 | 7168  | 1013        | 942          | 888          | 853         | 868          | 868          | 0.84         | 0.92        | 0.98         |
| 8192 | 8192 | 8192  | 979         | 897          | 861          | 852         | 853          | 870          | 0.87         | 0.95        | 1.01         |
| 4864 | 4096 | 8192  | 1123        | 1036         | 968          | 876         | 970          | 985          | 0.78         | 0.94        | 1.02         |
| 9728 | 8192 | 65536 | 1225        | 1075         | 968          | 964         | 968          | 1005         | 0.79         | 0.90        | 1.04         |
|      |      |       |             |              |              | 985         | 986          | 953          | NA           | NA          | NA           |


Sources of speedup / slowdown

1) For 16-bit GEMMs, added new configs.
2) For 8-bit GEMMs, slowdowns are due to adding of scale. Int8 slowdown unknown. Will investigate later. Currently we need this functionality more than absolute performance.